+++ src/main/kotlin/com/williambl/gameObject/GameObject.kt
@@ -2,11 +2,11 @@ package com.williambl.gameObject
 
 import com.williambl.renderObject.RenderObject
 import main.kotlin.com.williambl.Engine
-import org.joml.Vector3d
+import org.joml.Vector3f
 
 abstract class GameObject {
 
     abstract val engine : Engine
     abstract val renderObj : RenderObject?
-    abstract var position : Vector3d
+    abstract var position : Vector3f
 }
